typography moved back to catchup with git
